% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ograph_function.R
\name{buildLevels}
\alias{buildLevels}
\title{buildLevels}
\usage{
levels<-buildLevels(graph)
}
\arguments{
\item{graph}{an igraph object.}
}
\value{
a list object contains:
    \item{\code{nodes2level}:}{Object of class \code{"character"} ~~ }
    \item{\code{level2nodes}:}{Object of class \code{"character"} ~~ }
    \item{\code{noOfLevels}:}{Object of class \code{"character"} ~~ }
    \item{\code{noOfNodes}:}{Object of class \code{"ANY"} ~~ }
}
\description{
build a set of object to represent the igraph structure
}
\details{
build a set of object to represent the igraph structure
}
\author{
Xin he
}
\keyword{buildLevels}

