% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ograph_function.R
\name{buildOntGraph}
\alias{buildOntGraph}
\title{buildOntGraph}
\usage{
buildOntGraph(toTable(ONTTERM),toTable(ONTPARENTS))
}
\arguments{
\item{edges}{a data frame contain the relationship between nodes.
This is usually generate by using the ONTPARENTS object from the
topOnto.xxx.db package.}

\item{nodes}{a data frame contain the nodes and their definition.
This is usually generate by using the ONTTERM object from the
topOnto.xxx.db package.}
}
\value{
an igrph object represent the ontology
}
\description{
build the ontology with igraph.
Need to use topOnto.xxx.db
}
\details{
build the igraph object of the ontology
}
\author{
Xin he
}
\keyword{buildOntGraph}

