% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ograph_annotation.R
\name{mapGene2Graph}
\alias{mapGene2Graph}
\alias{readMappings}
\title{mapGene2Graph}
\usage{
...
}
\arguments{
\item{graph}{igraph object represent the ontology graph}

\item{file}{annotation file}
}
\description{
read annotation file and add a 'genes' attribute to graph node
}
\author{
Xin He
}

